requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv

passlib[bcrypt]
bcrypt

python-jose[cryptography]

authlib

itsdangerous